blend: added bias & MAE weight calculation
Mostly ad hoc code that follows our _current_ requirements. If a more
general solution is required, it will be addressed at that point.

Each of the different stages (bias, mae, blend) are meant to be executed
separately. 'raw' grids are interpolated and stored as a separate task to
the common grid.

To be tuned!