Optimize cape 500m interpolation by pre-sampling the data.
cape 500m source data is an average of the values found in the
lowest 500 meters. The source data is interpolated in 2 hPa intervals
with modifier. This is a very time consuming operation, and instead
the data is now sampled to correct intervals in one operation which
is a lot faster.

The downside of the approach is that when we are using linspace()
the partitions rarely exactly match the range we want, which means
that the last partition in this case is less than 2 hPa and is not
included in the averaging. Modifier does take this into account. What
we can do to actually improve the accuracy is to decrease the partition
length from 2 hPa to 1 hPa, which should make the overall value more
accurate even if we're disregarding the last interval.