blend: fetch bias & mae grids from historical data
Current method uses Radon directly to check if the wanted files exist,
since we don't want to fill the log with unwanted 'File not found'
messages. This method is quite slow in the worst case, making many
queries to the database (iteratively...). These queries could be
coalesced. Another option would be to add a new function to fetcher
that would allow start and end times to be specified.