HIMAN-236: Making info class a template.
Inside info class data is held in struct 'base' which is actually
the place where templatization starts, but as a result info class
is also transformed to a template.

This commit does not change Himan behavior as rest of Himan codebase
is still hard coded to use 'double' as template data type.