HIMAN-255: Most unstable CAPE improvements
Currently cape-plugin extracts the level that produces highest
theta e values (under 600hPa) and calculates mucape based on that.

On some occasions this procedure fails, and mucape is actually
found from some other level.

Because calculating mucape from all hybrid levels separately is
too expensive, use a less-compute intensive approach.

Instead of using just one max theta e level, create a vertical profile
of theta e values and find the local maxima(s) of the profile. The number
of maxima could be limited to say three. Then process mucape based
on the levels of the local maxima and choose the that produces
highest cape.

The upper limit of theta e search has been raised to 550 hPa, and the
maxima search limit has been defined at 650 hPa. This means that
theata e values are calculated up to the higher limit, but the maxima
search is stopped before that limit. Reason for this is that we want
to remove the artifical maxima value at 600hPa which is actually a
result of the search capping.

Includes also optimization for memory usage of gpu version of GetLFC()
and GetCAPE().