HIMAN-257: optimizing preform_hybrid
preform_hybrid examines the properties of the zero levels below 10km.
Initial implementation follow smarttool logic where all levels were
processed separately. This is not efficient when grid size is large.

This commit adds modifier the functionality that it returns all found
values in the given height range. This is triggered with 'findNth' value
-1. Default behavior is not changed and test results should not be
affected.