STU-9054: snow_drift plugin
Calculate snow drift value based on algorithm by
SNAPS/Vegsyn.

Algorithm is implemented as a plugin (instead of
a script) because it needs to be calculated sequentally
in time.

Output parameter is snow drift index, having values 0,1,2 or 3
(no drift, low, moderate, high drift). Plugin also writes two
intermediate parameters that are only used in the calculation
and are not supposed to be used directly.

When calculating snow drift forecast, we should 'calibrate' the
calculation using observed data. Therefore this plugin needs to
calculated to both observation producer (eg. LAPS) and to actual
forecast producer.