HIMAN-254: lower LFC search maximum height
LFC search is now capped at roughly 250hPa, as CAPE is not bound
to be found high in the atmosphere. Even 250hPa is quite high.

By lowering the maximum height we can skip reading 33 hybrid levels
of data (ECMWF) which is not a small amount. The change has zero
effect on the CAPE values; some LFC are changed though (ie. not found
anymore).