STU-9661: mask sea areas based on ice coverage data
In order to prevent snow drift calculation on open sea, ice coverage
data is read and calculation is done only if ice coverage >= 70%.
If ice coverage data is not found, land sea mask will be used with
threshold 0.5 (now integrated to this plugin, previously this was
done with a transformer call). If land sea mask data is not found,
all grid points are processed.