Prevent multiple threads from reading the same data simultaneously from disk
In some cases multiple threads can start to fetch the same data from
database/disk at the same time. This wastes cpu/disk resources and is unnecessary.

Prevent this race-condition fetch by synchronizing access. Each thread will
record their intent with a boost::shared_mutex. Multiple shared_mutexes
are not blocking each other. If a totally new grid is fetched (which is
not in cache), the first thread will upgrade the shared_mutex to unique_mutex,
which will block the other threads. Once fetch is complete, unique_mutex is
destroyed and the other (waiting) threads can simply pull the data from cache.

Note: the functionality is only enabled if data is unpacked directly after read,
because in the other case cache is not updated until later when the processing task
unpacks the data. This means that for a large (largest?) fraction of operation
Himan execution this change does not have any effect.