Add time offset to aggregation, remove GRIB1 support for sub-hour data
Add time offset to aggregation class which describes what it the starting
time of the aggregation period relative to current step. Usually offset is
the same as duration but with negative sign.

At the same time remove support for writing sub-hour data in GRIB1. eccodes
command line tools do not like this and GRIB2 has in any case much better
support for time metadata.