HIMAN-137: Add support for forecast type "statistical processing"
Add new forecast type "statistical processing" which is used for
outputs of fractile and probability plugins. This is required if/when
we change processing to be a new dimension in the database and remove it
from the parameter name. At that time we may have for example T-K which
could be from a deterministic forecast or from a statistical process.
We need the metadata to figure out which one it is.

Also this change is a nodge towards removing "producer type" information:
we should not restrain a single producer to just one forecast type. This
change is not finished yet as ecmwf data still needs some work.

Ref:STU-8537