HIMAN-290: Add support for writing to S3
Previous changes allow Himan to read from S3. Now writing is also possible.

Currently writing is only supported if write_mode = single.
This is because S3 does not support appending to a file (object).

Allowing Himan io-plugins (grib,querydata,...) to write mulitple messages in
a single pass requires a larger refactoring, which is a future task.