HIMAN-293: Allow filename to specified in configuration file
Filename template can be now set in the configuration file using
key 'filename_template'. This allows the name to be modified to suit
multiple purposes, instead of relying on the hard coded names.

For more information about the configuration options check
configuration.md.

boost::regex is used instead of std::regex because the regex support
is implemented for gcc version 4.9, and RHEL7 is using 4.8.5.

https://stackoverflow.com/questions/12530406/is-gcc-4-8-or-earlier-buggy-about-regular-expressions

$ g++ -v
[..]
gcc version 4.8.5 20150623 (Red Hat 4.8.5-39) (GCC)