HIMAN-305: Add geotiff support
Only reading is supported for now. Metadata is read from radon
database table "geotiff_metadata", and is matched to both file
metadata and (possible) band metadata.

Metadata can be a regular expression and if so boost::regex
is used to match the regex and the metadata.

Database data table column 'message_no' is used when fetching
band number.

A soft validation is done to metadata: Himan is comparing the
requested metadata to found metadata whenever it can ie. whenever
it is able to extract some piece of metadata from file.

Transformation functions are still mostly hard coded and a better way
should be found. Only for parameter name we do a database lookup
(table param_geotiff).