STU-14196: reduce gpu memory usage with vector component rotation
To properly rotate the vector components of a projected domain we need
to know the longitude of all the rotated points. The list of longitudes is
copied to gpu memory and passed to rotating function, one copy for each
processing thread. As the longitude values are the same for each thread,
all copies are equal.

Copying is changed so that for each distinct domain only one copy of the
longitudes is stored to gpu memory, and all threads share the same
(read only) data.