Battling my way through aggregation and processing type
Aggregation and processing type in grib2 are scattered in different
sections, and a problematic thing is that the key most often used
to define a statistical method 'typeOfStatisticalProcessing' is not
defined for instantaneous parameters. Therefore we have to bend
backwards a bit in order to read and write grib2 metadata correctly.

Also himan::aggregation is heavily geared towards time-aggregation.
Granted, this is the most common dimension to do aggregation. Aggregation
on other dimensions cannot be described very well currently (a TODO
item that one).