HIMAN-320: Create 500m cape source data with GPU
Performance profiling shows that creating the sampled 500m
cape source data is time consuming. It is also done with CPU,
even if other code is executed at GPU.

Translate the sampling code to GPU. Initial tests show
significant performance improvement and that peak memory consumption
happens in cape integration (ie. after sampling is done) so
this should be safe to put to production.