HIMAN-328: Fix serialization functions
Serialization code was broken as it is not compiled by default.
Several things have changed in the grid class structure which made
fixing the code almost like starting from the beginning.

grid-classes are immutable objects with no default constructor. The
parameter-constructors of each leaf class are initializing the members
of the parent class in their own required way. For a while this seemed
like a show stopper for cereal, but in the end the serialization does
work. The key takeaway is that the grid leaf classes do not serialize
the parent class, but use the (parent) class member variables to save
and load state.