Implement a simple cache for grid definitions
Initializing a new grid (area) using gdal has a fixed startup cost.
That is not much in absolute terms, but when called multiple times
(thousands and thousands) the overhead is accumulating.

Add a simple cache for grids that can be used to clone a new grid when
needed. Cloning is cheaper than creating a new grid from scratch. The
cache key is a string created from area information.

Currently functionality is used only in grib-plugin.