Add support for automatically fetching data from older forecasts Add more flexible Fetch function for lua scripts
In many cases if some data is not found, we want to use the previous
forecast data. Until now this had to be separately coded to each
plugin / lua script. Now a new option for fetcher exists which
will enable fetcher to automatically try to use older forecast if the
newest does not provide the requested data. By default this behavior
is switched off. The argument is a boolean flag called
'readFromPreviousForecastIfNotFound'.

To support this properly in lua scripts a new pair of fetch functions
are added: FetchWithArgs and FetchInfoWithArgs. These take as argument
a lua table which has the mandatory and optional fields as table keys.
The values should given as correct datatypes, for example for "param" use
object param instead of a string representation of the parameter name.

Mandatory fields are:
* forecast_time
* level
* param

Optional fields are:
* forecast_type, default: deterministic
* return_packed_data, false
* read_previous_forecast_if_not_found, false
* geom_name, use configuration value
* producer, use configuration value
* lsm_threshold, false
* do_interpolation, true
* do_level_transform, true
* do_vector_rotation, true
* use_cache, true