HIMAN-349: Spill write to disks that don't fit to memory
If a cache limit is set Himan will try to evict data from cache
if the limit is reached. In certain circumstances eviction
cannot be made for data that is not yet written, for example
because appending to an s3 object should be done
(which is impossible).

In this scenario data must be held in cache until the end of
Himan execution, which might mean that cache will grow to
larger than what is set as the upper limit. To prevent this
"spill" writes to disk.

This means that data that is to be written to s3 is serialized
to a file, and read back when it's time to write it. The data is
removed from disk afterwards. This means of course that necessary
disk space should be available.

Default directory for these spill files is /tmp, but can be
change with env variable HIMAN_TEMP_DIRECTORY.