Bind class 'params' ie. std::vector<param>
In many luascripts it's necessary to try to fetch
data with multiple names, for example N-0TO1 or N-PRCNT.

This is made easy with Himan c++ code, but no such effort
has been made for luatools.

With this commit it's possible to fetch from multiple
parameters with FetchWithArgs() function.

See total-cloud-cover.lua for examples.

TODO: the number of different Fetch() functions should
be cut down.