HIMAN-357: Add grid size to cache unique label
When a single configuration has multiple target_geom_names
(supported but not encouraged), there has to be a ways to
separate cache items with different geometries. Himan will
accept the cache item as-is, meaning that if geometry information
is not included, a grid might be returned that does not match
what was actually requested.